+++ jax/_src/mesh_utils.py
@@ -460,8 +460,8 @@ def _get_prime_factors(x: int) -> list[int]:
       x //= p
     if x == 1:
       return factors
-  else:
-    return [x]  # x is a prime number.
+  factors.append(x)  # remaining x is a prime larger than isqrt of the input.
+  return factors
 
 
 def _enumerate_feasible_logical_axis_assignments(